+++ CHANGELOG.rst
@@ -64,6 +64,7 @@ Added
     - Custom display for "external links" fields (#349)
     - Settings value ``SHEETS_EXTERNAL_LINK_LABELS`` (#349)
     - Custom display for different "contact" fields
+    - Handle sheet table cell overflow
 
 Changed
 -------